+++ src/lib/components/Card/Card.tsx
@@ -16,7 +16,7 @@ export const Card = (props: ICardProps) => {
 		const cssClasses: Array<string> = [];
 		cssClasses.push(styles.card);
 		shadow && cssClasses.push(styles.shadow);
-		cssClasses.push(className);
+		className && cssClasses.push(className);
 		return cssClasses.filter(css => css).join(' ');
 	};