+++ src/lib/components/Card/CardImage.tsx
@@ -10,7 +10,7 @@ export const CardImage = ({ src, alt, className = '', ...rest }: ICardImageProps
 	const getCssClasses = () => {
 		const cssClasses: Array<string> = [];
 		cssClasses.push('card-img-top');
-		cssClasses.push(className);
+		className && cssClasses.push(className);
 		return cssClasses.filter(css => css).join(' ');
 	};