+++ src/lib/components/Checkbox/Checkbox.tsx
@@ -44,7 +44,7 @@ export const Checkbox = (props: ICheckboxProps) => {
 	const getCssClasses = () => {
 		const cssClasses: Array<string> = [];
 		cssClasses.push('checkbox');
-		cssClasses.push(className);
+		className && cssClasses.push(className);
 		return cssClasses.filter(css => css).join(' ');
 	};