+++ src/lib/components/Grid/Column/Column.tsx
@@ -26,13 +26,13 @@ export const Column = (props: IColProps) => {
 
 	const getCssClasses = () => {
 		const cssClasses: Array<string> = [];
-		cssClasses.push(className);
 		!xs && !sm && !md && !lg && !xl && cssClasses.push('col');
 		if (xs) cssClasses.push(`col-${xs.toString()}`);
 		if (sm) cssClasses.push(`col-sm-${sm.toString()}`);
 		if (md) cssClasses.push(`col-md-${md.toString()}`);
 		if (lg) cssClasses.push(`col-lg-${lg.toString()}`);
 		if (xl) cssClasses.push(`col-xl-${xl.toString()}`);
+		className && cssClasses.push(className);
 		return cssClasses.filter(css => css).join(' ');
 	};