+++ src/lib/components/Grid/Row/Row.tsx
@@ -11,8 +11,8 @@ export const Row = (props: IRowProps) => {
 
 	const getCssClasses = () => {
 		const cssClasses: Array<string> = [];
-		cssClasses.push(className);
 		cssClasses.push('row');
+		className && cssClasses.push(className);
 		return cssClasses.filter(css => css).join(' ');
 	};