+++ src/lib/components/Icon/Icon.tsx
@@ -12,7 +12,7 @@ export const Icon = (props: IIconProp) => {
 	const getCssClasses = () => {
 		const cssClasses: Array<string> = [];
 		cssClasses.push(styles.icon);
-		cssClasses.push(className);
+		className && cssClasses.push(className);
 		return cssClasses.filter(css => css).join(' ');
 	}