+++ src/lib/components/SpeedDial/SpeedDialIcon.tsx
@@ -11,7 +11,6 @@ export const SpeedDialIcon = (props: ISpeedDialIconProps) => {
 
 	const getCssClasses = () => {
 		const cssClasses: Array<string> = [];
-		// cssClasses.push(styles.speedDial);
 		className && cssClasses.push(className);
 		return cssClasses.filter(css => css).join(' ');
 	};