+++ src/lib/components/Tabs/Tabs.tsx
@@ -31,7 +31,7 @@ export const Tabs = (props: ITabsProps) => {
 	const getCssClasses = () => {
 		const cssClasses: Array<string> = [];
 		cssClasses.push(styles.tabs);
-		cssClasses.push(className);
+		className && cssClasses.push(className);
 		return cssClasses.filter(css => css).join(' ');
 	};