+++ src/lib/components/index.ts
@@ -2,6 +2,7 @@ export * from './AppBar';
 export * from './Backdrop';
 export * from './Badge';
 export * from './Button';
+export * from './ButtonGroup';
 export * from './Breadcrumb';
 export * from './Card';
 export * from './ConditionalWrapper';